Change all voters to submit USER_BLOCKING votes
The ProcessPriorityPolicy class eventually merges all USER_VISIBLE and
USER_BLOCKING votes into a single `is foreground` bit, meaning they have
the same behavior.

In the future, we want to experiment with multiple process priorities.
To do so, the first step is to consolidate all TaskPriority enum
usage to the same value so that other values can have a different
behavior in the future.

Change-Id: I9c59dc01a7c28e8b3fdfd141123bf3fee762eea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5709650
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328253}