[views-ax] Fix cached role/name/description out of sync in the cache (//chrome/views)
We just discovered that CL:5537333 introduced an unexpected change in
behavior for a subset of views around the accessible role, name, and
description. The change is around what role, name, or description to
apply: the one in the ViewAccessibility cache, or the one computed in
the View::GetAccessibleNodeData override. Prior to CL:5537333, the
latter was exposed, but since then, it's the one in the cache that has
the final say.

Ultimately, when the ViewsAX refactor is complete, we'll only ever
expose what is in the ViewAccessibility cache and we'll get rid of
View::GetAccessibleNodeData. However, today, we still care about
maintaining the right ordering between both systems to expose the
right value.

Before we landed CL:5537333, we carefully fixed all views whose tests
indicated a regression. However, several more views did not have the
right accessible tests in place (or did not have any) to prevent such
a regression.

While no user facing bug has been reported thus far in Canary or Beta,
we decided it is safer to fix all potential regressions we introduced
in that change. It's possible that some views are currently exposed
with a button role instead of a toggle button role, for example, and
that the accessible name/description of a bookmark doesn't change
after it gets renamed. Again, we are not aware of any user facing bug
at the moment, but it doesn't mean that there aren't any -- we have a
small population of AT users in Canary and Beta.

The issue happens when a base class sets its role, name, or
description in the cache through a setter, while a subclass sets it
in a View::GetAccessibleNodeData override. The value set in the
override should be the one exposed, but the one set on the base class
ends up being the one that does.

Here's an example:
* InMenuButton should be exposed as a menu item, so it sets its role
  to kMenuItem in GetAccessibleNodeData.
* InMenuButton extends LabelButton, which extends Button.
* Button sets its role to kButton directly in the ViewAccessibility
  cache.
* The role exposed for InMenuButton is kButton instead of kMenuItem.

This CL fixes all instances of this issue in //chrome/browser. Only
one non-//chrome/browser file needs to be modified in this class: ui/views/controls/button/menu_button_controller.cc. This is necessary
because the menu button controller is used on several
//chrome/browser views that are subclasses of the MenuButton.

Given that CL:5537333 is in 127 and 127's Stable cut is on 7/16, we
did those fixes quickly without adding new tests for now. However,
we plan to add tests in a follow up CL.

Bug: 352528963, 325137417
Change-Id: I3b95a9803d5bc334b41a8884e8a1266025452e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5689427
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Auto-Submit: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328263}