MessagePumpEpoll: Fix to disable epoll events properly
The current implementation of MessagePumpEpoll calls
epoll_ctl(EPOLL_CTL_MOD) with events == 0 if there is no active interest
in order to disable the fd, but this may cause failing to handle the
EPOLLHUP event and end up removing the fd from the interests of epoll
unintentionally.
This CL fix to disable the fd in a more proper way.

./base/message_loop/message_pump_libevent_unittest.cc`
./base/files/file_descriptor_watcher_posix_unittest.cc`

Bug: 349729214
Test: `./tools/autotest.py -C out/Default
Test: `./tools/autotest.py -C out/Default
Change-Id: I6f1cd344c08eb972bdcc9a8fcfccee9a5ed43200
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660161
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Olivier Li <olivierli@chromium.org>
Commit-Queue: Olivier Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328269}