webnn: Use custom allocations to avoid copies in TFLite compute()
This is a blocker for the dispatch() implementation which will also use
custom allocations. TFLite doesn't support removing a tensor from the
list of tensors that are using custom allocations so we need to have
compute() use them or else switching between compute() and dispatch()
will cause errors. That we can avoid copies as long as the BigBuffer is
properly aligned is a nice bonus.

Change-Id: I30803a2b3eaa7c8436d04e8a6bf0a743d979997b
Bug: 352610625
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5709509
Reviewed-by: Phillis Tang <phillis@chromium.org>
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Phillis Tang <phillis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328271}