Check whether a group-controlled feature is enabled for a profile.
Google groups controlled studies are studies that have a google_groups
parameter (or a exclude_google_groups parameter) in their filter. The
google_groups parameter works as follows:
If any of the installation's profiles are part of at least one of the
groups specified in the google_groups parameter, then the filter is
fulfilled. If the study's chosen experiment enables a feature, then
that feature is enabled for the entire installation - regardless of the
profile that is currently active. Put differently, since
base::FeatureList::IsEnabled is only a function of a base::Feature and
not a profile, the feature state is global to the entire Chrome
installation.

This CL adds the following functionality: If a feature is enabled by a
study with a google_groups parameter in its filter, you can now check
whether the feature would also be enabled if only the current profile
was installed.

Bug: 348575889
Change-Id: I423fff31046a3c9bd2250e1e606345c754314689
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5691433
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1328318}