[ios] Fix for PushNotificationAccountContextManager/removeAccount
This is a partial fix for an issue where removeAccount is called when
the account was never added. The DCHECK is left in place for now, but
the method will now return NO to indicate that the account was not
removed.

Fixed: 349858707
Change-Id: Ic316d8ed1165a46698ec3b2dc371d8ba9a1c99c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5713773
Reviewed-by: Daniel White <danieltwhite@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1328341}