[declutter] Plumb through and use rootView for accessibility logic
There's existing logic in TabListEditorLayout to deactivate
accessibility for obscured elements when the view is shown. Extending
that to support custom views which embed TabListEditor by plumbing
through a rootView argument. This argument is used in place of the
parent view.

Also add an lifecycle observer to TabListEditor which allows the
archived tabs dialog to be notified when the TabListEditor hides
itself. This will be used to (1) remove the embedding view from
the hierarchy to comply with TabListEditor expectations and (2)
be notified if the TabListEditor hides itself outside the dialog
control flow. This will eliminate any issues with the embedding
view (e.g. the close all tabs button) sticking around after the
TabListEditor is closed.

Bug: 348590015
Change-Id: I13b9e273eac38735344113eca25410c5fa2eeec6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5709795
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Brandon Wylie <wylieb@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1328356}