Add attribution fields to collaboration metadata in SyncEntity.
To be able to understand who made a particular change (e.g. creation
or updates) we need to store the data somewhere in Chrome Sync.

This CL adds this to the CollaborationMetadata message in the
SyncEntity proto, but does not yet expose it to any C++ structs.

Bug: 352214654
Change-Id: I4cffec0f1021b5869760e626818f1b3cf8f5ffc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5691270
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328365}