[iOSKAUpgrade] Fix expanded view horizontal misalignment
In the new expanded view, the header view (composed of the segmented
control, the Chrome icon and the close button) should be horizontally
aligned with the table view cells under.

Since the horizontal insets of the table view cells are determined by
the system and vary depending on the screen size, setting the header
view horizontal constraints with an arbitrary constant
(`kHeaderViewHorizontalPadding`), led to a misalignment on some
devices.

This CL fixes the issue by applying the cells' horizontal insets to
the header view. This CL also updates the default inset value from 20
to 16px per design.

Before/After comparison on iPhone SE:
https://screenshot.googleplex.com/4pbwxjxTqHwmHn7.png

Fixed: 350079646
Change-Id: Ifc69b22efb9ca9a63fffa2bdd40d966d5aa35258
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5714331
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Reviewed-by: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1328393}