[TabGroupSync] Test the wrapper service in both flows
Tests that the TabGroupServiceWrapper performs as expected in both flows
(SavedTabGroupKeyedService and TabGroupSyncService).

This is needed in preparation for the Migration from STGKS to TGSS. This way we will have test coverage after we have fully migrated over. Additionally, we will know if something breaks with one of the paths as we make changes over time.

Also implements AddGroup and RemoveGroup in the TabGroupServiceWrapper.

Change-Id: Ie62896b67a90dd0395ef4180b034eefff5993d09
Bug: 350514491
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5714869
Commit-Queue: Darryl James <dljames@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1328406}