[Tab Grid Dialog] Move share button to toolbar
Move the share button from the bottom share bar (soon to be deprecated)
to the new 2-row top toolbar. The style is very different, but the
functionality is unchanged. The legacy code will be removed in a
follow-up CL.

Bug: 347981599
Change-Id: Iaa7e0204568badb486aeecc78e515d8d12917ed5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5707492
Reviewed-by: Sky Malice <skym@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328455}