birch: Fix dangling pointer in birch bar view
This can happen while dragging the split view divider, which causes
a regular overview exit, which moves ownership of the birch bar to
the controller. Then the root window gets shutdown, but the birch bar
still has a pointer to it.

This is not the birch bar view's issue, but the root window pointer is
not needed yet.

Test: none
Bug: b/349164737
Change-Id: I659e462143671295a69cbb724fcb81d24498a809
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5713833
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328459}