[PDF Ink Signatures] Add brush size options in side panel UI
Add size options to the annotation mode side panel UI so that the user
can select the brush size.

When a user chooses between brushes, the size setting should be saved
in the current session. When the user goes back to a brush they've
already used, it will be the same size that the user last set.

Icons and actual brush sizes are temporary and need to be set in the
future. Styling will be added in the future.

Note that Ink2 also supports using the original Ink's annotation bar.
The size values are different and are scaled in the PDF content. The
side panel only supports Ink2, so the size values can be used as-is
without scaling. Stop scaling the size. This leaves Ink's annotation bar
sizing in a broken state, but it will be deprecated soon.

Bug: 351868764
Change-Id: I4adb9915ef204b67e39bf34e6e43e7f2484ca92b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5690924
Commit-Queue: Andy Phan <andyphan@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328490}