[Extensions] Add re-enable extension dialog for MV2 disabled stage
Chrome has announced that MV2 extensions will gradually no longer be
supported in the browser. None of these changes will take effect on any
Chrome channel before June, 2024. See more information at
https://developer.chrome.com/blog/resuming-the-transition-to-mv3.

The first stage of this is a user-visible notice in chrome://extensions,
which can be acknowledged. The second stage will disable unsupported MV2
extensions and will also have a user-visible notice.

This CL adds a dialog that will be shown when the user decided to
re-enable an extension disabled due to the MV2 deprecation. Dialog
will be connected to the trigger in the next CL.

Screenshot:
https://drive.google.com/file/d/134TzYzGxb8egF3_iI_3JDLsd-MxfopeQ/view?usp=sharing

Bug: 339061151
Change-Id: Ib90036a495a6a7fb4da968f368ed58ba53ba2968
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5714952
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328524}