CrOS WebUI: Use assert helpers from chai_assert.js, part 3.
This is in preparation of rolling Chai to v5, which breaks js2gtest
usage of Chai, since in v5 it is now only available as an ES Module and
not as a global 'window.chai' object.

Updating some CrOS usages in an effort to possibly avoid having to keep
Chai v4 (side-by-side with v5) for remaining js2gtest users.

Bug: 347991521
Change-Id: I62e1f8a13e7ce24ba697b2384d9c4f544d64cb23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5709180
Auto-Submit: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Charles Zhao <charleszhao@chromium.org>
Reviewed-by: Charles Zhao <charleszhao@chromium.org>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328530}