[Extensions] Handle global notice acknowledgment of MV2 disabled stage
Chrome has announced that MV2 extensions will gradually no longer be
supported in the browser. None of these changes will take effect on any
Chrome channel before June, 2024. See more information at
https://developer.chrome.com/blog/resuming-the-transition-to-mv3.

The first stage of this is a user-visible notice in chrome://extensions,
which can be acknowledged. The second stage will disable unsupported MV2
extensions and will also have a user-visible notice.

This CL connects the dismiss button in the second stage with the
actual dismissal of the MV2 deprecation notice. Once the notice is
dismissed, it's hidden for the rest of the experiment stage.

Screencast:
https://drive.google.com/file/d/1X77KGMjcoNi0_XQXKoIb0-d_rlF4GGZ-/view?usp=sharing

Bug: 339061151
Change-Id: I0a7ac641314dbedf686911601e4baa37c19d379f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5703552
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328550}