[Autofill] Add AutofillAgent::GetDocument()
AutofillAgent currently calls in many places
  unsafe_render_frame()->GetWebFrame()->GetDocument()
and after a followup CL, FormCache would need to do the same.

Therefore, this CL adds the shorthand AutofillAgent::GetDocument().

Unlike unsafe_render_frame(), the function is not prefixed with
"unsafe_" because WebDocument is refcounted, so keeping references
to WebDocument is fine.

The CL also replaces several early guards of the form
  if (!unsafe_render_frame()) return;
or similar with something equivalent to
  if (!GetDocument()) return;
This is safe because none of those are followed by unguarded
dereferences of unsafe_render_frame().

Finally, it fixes a potential null dereferences of
`element.GetDocument().GetFrame()`.

Bug: 40100455
Change-Id: I54dda7edaf719243e440f60a65902bfd39f0f951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5713655
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1328656}