Extensions: Ensure install dialog is positioned above the line of death
If the user starts extension installation from Chrome Web Store and quickly
moves to another tab, the permissions dialog is positioned below the line
of death.

The dialog must be parented to the top-level browser window to ensure correct
positioning. WebContents::GetTopLevelNativeWindow() does not return the top
level window if the WebContents are not active. It returns the inner native
window instead, which results in incorrect positioning. Detailed explanation
can be found in the linked bug.

This change activates the tab which initiated the installation before
showing the prompt. This ensures that GetTopLevelNativeWindow() returns
the correct window.

This also required a refactor in ExtensionInstallPromptShowParams because
it originally cached the value of WebContents::GetTopLevelNativeWindow()
at construct time. Thus, the correct window was not being returned even
after switching to the original tab.

Bug: 40059470
Change-Id: I0119edc0af8ccf17e0f16c1080a2920961ef2584
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5688271
Commit-Queue: Sohail Rajdev <sorajdev@microsoft.com>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328659}