Handle slow attempt and stagger IPv4/IPv6 in HttpStreamPool
This CL implements the "connection attempt delay" [1], which is used
to start the next connection attempt when the previous attempt is
considered slow. HttpStreamPool::Job first prefers to use an IPv6
endpoint (if available). When the connection attempt delay timer is
fired, Job tries to pick IPEndPoint from the different address
family (if the previous attempt used IPv6, it prefers to use IPv4).

[1] https://datatracker.ietf.org/doc/html/rfc8305#section-5

Bug: 346835898
Change-Id: I761d8d92bfb613bb0b153282fd4b3f41f4a2363d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5711190
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328664}