[ios] Make ServiceIsCreatedWithBrowserState() safe
There is some ordering issues between the initialisation of the
ChromeBrowserStateImplIOData, PrefService and some KeyedService.

This was causing the ServiceIsCreatedWithBrowserState() method
to be tricky on iOS as it could lead to crash if any of the
dependent KeyedService needed to call CreateRequestContext().

To fix this, split the initialisation of the DependencyManager.
First it is informed of the creation of the BrowserState, thus
allowing to create KeyedService, then initialisation of the
BrowserState is resumed, and finally the DependencyManager is
told to create the services.

Fixed: 342112724
Change-Id: I402b3b81568867ad0f2f272aded005b2610bb0e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5558505
Reviewed-by: Colin Blundell <blundell@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328737}