SearchPrefetch: Adjust the size of cache entries depending on "cache_size"
SearchPrefetchService manages a map from search URL to prefetch URL, so
that subsequent search requests could load prefetched responses from the
disk cache (not from the in-memory cache). This map is stored in the
prefs. The size of the map is capped by the "cache_size" field trial
parameter.

Before this CL, the map size was checked only when a new entry was
added, and the map never shrunk even if the "cache_size" is lowered
through the field trial parameter (Exception: the cache is cleared when
a user requests to clear browsing data or the template URL service is
changed).

To fix it, this CL checks the "cache_size" when the map is loaded from
the prefs, and discards extra entries if needed.

This change is necessary for experimenting on various "cache_size".

Bug: 353628436
Change-Id: I2ce464897e12e8a2bd4a03648e162cafc4e8a244
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5715896
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328753}