Fix the eye icon in the SaveUpdate bubble
This CL reverts the changes that were done to the layout manager as
part of a bigger cleanup. That is needed to restore the full
functionality of the eye icon that shows the plaintext password value.

The animated layout manager is probably not needed, but to make sure
that the this fix lands before the branch point, let's go back to
the code that we knew that was working.

Bug: 353552550
Change-Id: I39071c646667f593f9aefb22a811e47cf81b25bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5714830
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328762}