[iOS] Add Inactive Tabs button cell
The button is displayed in the regular grid, so it should be with the
regular grid, not with inactive tabs.

This CL adds the inactive tabs button cell. It also moves the
accessibility constant to "grid".
The layout still need to be updated to have the correct height.

https://screenshot.googleplex.com/6Ej9KhaZ9rM6G7s

Low-Coverage-Reason: TESTS_IN_SEPARATE_CL refactor of existing feature
Bug: 352722446
Change-Id: Ifd9db584a2167a6ab49a160de5aa1317e055b411
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5713154
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1328873}