Create a separate enum for CSS math expression keyword context
Relative colors based on currentcolor can't be resolved at parse time.
When such a color has a calc expression for one of its channels, we'll
need to build and store a representation that includes placeholders
for channel keywords in the expression. Currently, for similar cases,
we use the CSSMathOperator enum, but for color channel expressions
there is no associated operator.

This CL factors the concept of "context" for a keyword in a calc
expression to its own enum to pave the way for adding a context for
color channel keywords.
 

Bug: 325309578
Change-Id: Id9903b9f68d4ca74ca506ca57698a5c43d38c2d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5716348
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328900}