[iOS] Add iPad cursor interactions for Contextual Panel views
This CL adds extra iPad cursor interactions to the Contextual Panel,
making the experience much better for iPad users.

Beyond standard button interactions, this also adds an onContinuousHover
to the price graph, so the tooltip will appear when hovering the graph
and not just when dragging on the graph.

Bug: 352605064
Change-Id: I2cfde33f74dbbc104a9e212e43bfb54268cc0c46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5698470
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Nicolas MacBeth <nicolasmacbeth@google.com>
Reviewed-by: Cheick Cisse <cheickcisse@google.com>
Cr-Commit-Position: refs/heads/main@{#1328922}